Compilation of closures now captures free vars
The initial pass of closures wrapped all functions in a closure object
with no free variables and emits a closure instruction w/ a hardcoded 0
free variable len. This commit updates the compiler adding support for
emitting instructions for our free variables and including the free var
length in the OpClosure instruction.